mymemcpy is used

+++ mystdlib/src/mymemcpy.c
@@ -1,6 +1,6 @@
 #include <mymemcpy.h>
 
-void *mymemcpy(void *dest, const void *src, size_t n) {
+void *mymemcpy(void *dest, const void *src, int n) {
 
     int i;
     for (i = 0; i < n; ++i) {
